3-audio-generator/requirements.txt
google-cloud-texttospeech==2.16.3
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
functions-framework==3.5.0
